+++ Tests/Unit/SharePointDsc/SharePointDsc.SPSearchServiceApp.Tests.ps1
@@ -43,6 +43,7 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
         Mock -CommandName Set-SPEnterpriseSearchServiceApplication -MockWith {} 
         Mock -CommandName New-SPBusinessDataCatalogServiceApplication -MockWith { }
         Mock -CommandName Set-SPEnterpriseSearchServiceApplication -MockWith { } 
+        Mock -CommandName Set-SPEnterpriseSearchService -MockWith {}
         
         Mock -CommandName Get-SPEnterpriseSearchServiceInstance -MockWith { 
             return @{} 
@@ -62,6 +63,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
         } -ParameterFilter { 
             $TypeName -eq "Microsoft.Office.Server.Search.Administration.Content" 
         }
+        Mock -CommandName Get-SPEnterpriseSearchService -MockWith {
+            return @{
+                ProcessIdentity = "DOMAIN\username"
+            }
+        }
         
         Mock Import-Module -MockWith {} -ParameterFilter { $_.Name -eq $ModuleName }
 
@@ -71,6 +77,7 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                 Name = "Search Service Application"
                 ApplicationPool = "SharePoint Search Services"
                 Ensure = "Present"
+                WindowsServiceAccount = $mockCredential
             }
 
             Mock -CommandName Get-SPServiceApplication -MockWith { 
@@ -525,6 +532,85 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                 { Set-TargetResource @testParams } | Should Throw
             }
         }
+
+        Context "A service app exists that has a correct windows service account in use" -Fixture {
+            $testParams = @{
+                Name = "Search Service Application"
+                ApplicationPool = "SharePoint Search Services"
+                Ensure = "Present"
+                WindowsServiceAccount = $mockCredential
+            }
+            
+            Mock -CommandName Get-SPServiceApplication -MockWith { 
+                $spServiceApp = [PSCustomObject]@{
+                    TypeName = "Search Service Application"
+                    DisplayName = $testParams.Name
+                    ApplicationPool = @{ Name = $testParams.ApplicationPool }
+                    Database = @{
+                        Name = $testParams.DatabaseName
+                        Server = @{ Name = $testParams.DatabaseServer }
+                    }
+                }
+                $spServiceApp = $spServiceApp | Add-Member -MemberType ScriptMethod -Name GetType -Value { 
+                    return @{ FullName = $getTypeFullName } 
+                } -PassThru -Force
+                return $spServiceApp
+            }
+
+            it "Should return the current value in the get method" {
+                (Get-TargetResource @testParams).WindowsServiceAccount | Should Not BeNullOrEmpty 
+            }
+
+            it "Should return true in the test method" {
+                Test-TargetResource @testParams | Should Be $true
+            }
+        }
+
+        Context "A service app exists that has an incorrect windows service account in use" -Fixture {
+            $testParams = @{
+                Name = "Search Service Application"
+                ApplicationPool = "SharePoint Search Services"
+                Ensure = "Present"
+                WindowsServiceAccount = $mockCredential
+            }
+            
+            Mock -CommandName Get-SPServiceApplication -MockWith { 
+                $spServiceApp = [PSCustomObject]@{
+                    TypeName = "Search Service Application"
+                    DisplayName = $testParams.Name
+                    ApplicationPool = @{ Name = $testParams.ApplicationPool }
+                    Database = @{
+                        Name = $testParams.DatabaseName
+                        Server = @{ Name = $testParams.DatabaseServer }
+                    }
+                }
+                $spServiceApp = $spServiceApp | Add-Member -MemberType ScriptMethod -Name GetType -Value { 
+                    return @{ FullName = $getTypeFullName } 
+                } -PassThru -Force
+                return $spServiceApp
+            }
+
+            Mock -CommandName Get-SPEnterpriseSearchService -MockWith {
+                return @{
+                    ProcessIdentity = "WrongUserName"
+                }
+            }
+            
+            it "Should return the current value in the get method" {
+                (Get-TargetResource @testParams).WindowsServiceAccount | Should Not BeNullOrEmpty
+            }
+
+            it "Should return false in the test method" {
+                Test-TargetResource @testParams | Should Be $false
+            }
+
+            it "Should update the account in the set method" {
+                Set-TargetResource @testParams
+
+                Assert-MockCalled -CommandName "Set-SPEnterpriseSearchService"
+            }
+        }
+
     }
 }